refactor: bring references and otus onto the query conventions
Replace useUpdateSourceTypes(key, path, queryKey, sourceTypes) with two
plain mutations in references/queries.ts — useUpdateReferenceSourceTypes
and useUpdateDefaultSourceTypes — each with its own path and invalidation.
The form and undo machinery it bundled is not query-layer, so it moves to
useSourceTypeEditor beside the two components that share it, leaving
references/hooks.ts with only the reference-rights helpers.

Drop the mutation-level onSuccess that recomputed lastRemoved. It was
dead: sourceTypes.length > updatedSourceTypes compares a number to an
array, coercing to NaN, so the branch never ran; and where it could run —
removing the last remaining type, where [] coerces to 0 — the per-call
onSuccess overwrote it with the same value.

Move the OTU context provider out to otus/components/CurrentOtuContext.tsx
so otus/queries.tsx can become a .ts file, put the cross-feature imports
on the @references/* alias, and fold getGenbank into useGetGenbank. A
button press drives that read rather than what is on screen, so there is
nothing to key it on and nothing worth caching, making a mutation the fit;
Accession sheds both its useState calls as a result.
