+++ apps/web/src/references/hooks.ts
@@ -1,119 +1,5 @@
 import { useFetchAccount } from "@account/account";
-import { type ApiResponse, apiClient } from "@app/api";
-import { zodResolver } from "@hookform/resolvers/zod";
-import {
-	type QueryKey,
-	useMutation,
-	useQueryClient,
-} from "@tanstack/react-query";
-import { difference, union } from "es-toolkit/array";
-import { useState } from "react";
-import { useForm } from "react-hook-form";
-import z from "zod";
 import { useFetchReference } from "./queries";
-
-export function getValidationSchema(sourceTypes: string[]) {
-	return z.object({
-		sourceType: z
-			.string()
-			.transform((val) => val.toLowerCase().trim())
-			.refine(
-				(val) => !val.includes(" "),
-				"Source types may not contain spaces",
-			)
-			.refine(
-				(val) => !sourceTypes.includes(val),
-				"Source type already exists",
-			),
-	});
-}
-
-export function useSourceTypesForm(sourceTypes: string[]) {
-	const {
-		formState: { errors },
-		handleSubmit,
-		register,
-		reset,
-	} = useForm({
-		defaultValues: { sourceType: "" },
-		resolver: zodResolver(getValidationSchema(sourceTypes)),
-	});
-
-	return { errors, handleSubmit, register, reset };
-}
-
-export function useUpdateSourceTypes(
-	key: "default_source_types" | "source_types",
-	path: string,
-	queryKey: QueryKey,
-	sourceTypes: string[],
-) {
-	const queryClient = useQueryClient();
-
-	const [lastRemoved, setLastRemoved] = useState("");
-
-	const mutation = useMutation({
-		mutationFn: async (sourceTypes: string[]): Promise<ApiResponse> => {
-			return apiClient.patch(path).send({ [key]: sourceTypes });
-		},
-		onSuccess: (data: ApiResponse) => {
-			const updatedSourceTypes = data.body[key];
-
-			if (sourceTypes.length > updatedSourceTypes) {
-				setLastRemoved(difference(sourceTypes, updatedSourceTypes)[0] ?? "");
-			} else {
-				setLastRemoved("");
-			}
-
-			queryClient.invalidateQueries({ queryKey });
-		},
-	});
-
-	const { errors, handleSubmit, register, reset } =
-		useSourceTypesForm(sourceTypes);
-
-	function handleAdd({ sourceType }) {
-		if (sourceType) {
-			mutation.mutate(union(sourceTypes, [sourceType]), {
-				onSuccess: () => {
-					reset();
-					setLastRemoved("");
-				},
-			});
-		}
-	}
-
-	function handleRemove(sourceType) {
-		mutation.mutate(
-			sourceTypes.filter((s) => s !== sourceType),
-			{
-				onSuccess: () => {
-					setLastRemoved(sourceType);
-				},
-			},
-		);
-	}
-
-	function handleUndo() {
-		if (lastRemoved) {
-			mutation.mutate(union(sourceTypes, [lastRemoved]), {
-				onSuccess: () => {
-					setLastRemoved("");
-				},
-			});
-		}
-	}
-
-	return {
-		error: errors.sourceType?.message,
-		lastRemoved,
-		mutation,
-		handleRemove,
-		handleSubmit: handleSubmit(handleAdd),
-		handleUndo,
-		register,
-	};
-}
 
 /**
  * Returns whether the reference is archived. `false` while the reference query is loading.
