+++ apps/web/src/sequences/components/SequenceButtons.tsx
@@ -1,7 +1,7 @@
 import CloseButton from "@base/CloseButton";
 import IconButton from "@base/IconButton";
+import { useCurrentOtuContext } from "@otus/components/CurrentOtuContext";
 import { useGetActiveIsolateId } from "@otus/hooks";
-import { useCurrentOtuContext } from "@otus/queries";
 import { DownloadLink } from "@references/components/Detail/DownloadLink";
 import {
 	useCheckReferenceRight,
